backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
mysqlclient
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
email-validator
python-multipart
openpyxl
pandas
requests
google-auth
ollama>=0.1.0



